+++ django_app/redbox_app/redbox_core/views/metrics_view.py
@@ -5,7 +5,7 @@ from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 
-from redbox.models import Settings
+from redbox import Settings
 
 logging.basicConfig(level=logging.ERROR)
 env = Settings()
