+++ redbox-core/.vscode/tasks.json
@@ -1,23 +1,5 @@
 
 {
     "version": "2.0.0",
-    "tasks": [
-        {
-            "label": "Generate Graphs",
-            "type": "shell",
-            "command": "poetry run python ../utilities/draw_graph.py",
-            "options": {
-                "env": {
-                    "MINIO_HOST": "localhost",
-                    "POSTGRES_HOST": "localhost",
-                    "ELASTIC__HOST": "localhost"
-                }
-            },
-            "presentation": {
-                "reveal": "always",
-                "panel": "new"
-            },
-            "problemMatcher": []
-        }        
-    ]
+    "tasks": []
   }
