--- redbox-core/redbox/__init__.py
@@ -1,7 +1,0 @@
-"""Redbox is a Python library for working with the Redbox API, data and services."""
-
-from redbox.app import Redbox
-
-__version__ = "0.3.0"
-
-__all__ = ["Redbox"]
