--- redbox-core/redbox/chains/components.py
@@ -1,8 +1,0 @@
-from functools import cache
-
-import tiktoken
-
-
-@cache
-def get_tokeniser() -> tiktoken.Encoding:
-    return tiktoken.get_encoding("cl100k_base")
