--- redbox-core/redbox/models/__init__.py
@@ -1,7 +1,0 @@
-from redbox.models.settings import Settings
-from redbox.models.chain import RedboxState
-
-__all__ = [
-    "RedboxState",
-    "Settings",
-]
