--- redbox-core/redbox/models/chain.py
@@ -1,37 +1,0 @@
-from langchain.chat_models import init_chat_model
-from langchain_core.documents import Document
-from langchain_core.messages import AnyMessage, BaseMessage
-from langchain_core.prompts import PromptTemplate
-from pydantic import BaseModel, Field
-
-from redbox.models.settings import ChatLLMBackend, get_settings
-
-
-class RedboxState(BaseModel):
-    documents: list[Document] = Field(description="List of files to process", default_factory=list)
-    messages: list[AnyMessage] = Field(description="All previous messages in chat", default_factory=list)
-    chat_backend: ChatLLMBackend = Field(description="User request AI settings", default_factory=ChatLLMBackend)
-
-    def get_llm(self):
-        if self.chat_backend.provider == "google_vertexai":
-            return init_chat_model(
-                model=self.chat_backend.name,
-                model_provider=self.chat_backend.provider,
-                location="europe-west1",
-                # europe-west1 = Belgium
-            )
-        return init_chat_model(
-            model=self.chat_backend.name,
-            model_provider=self.chat_backend.provider,
-        )
-
-    def get_messages(self) -> list[BaseMessage]:
-        settings = get_settings()
-
-        input_state = self.model_dump()
-        system_messages = (
-            PromptTemplate.from_template(settings.system_prompt_template, template_format="jinja2")
-            .invoke(input=input_state)
-            .to_messages()
-        )
-        return system_messages + self.messages
